refactor: Move zod-schema to shared-package
