+++ apps/core/src/services/usersService.ts
@@ -3,7 +3,7 @@ import { mapReviewStatus } from "../utils/enumMapper.js";
 import type {
   GetUserReviewsQuery,
   UpdateUserInfoBody,
-} from "../zod_schemas/user.schema.js";
+} from "@cugetreg/zod-schemas/user";
 
 export const usersService = {
   getUserInfo: async (userId: string) => {
