--- apps/core/src/zod_schemas/constants.ts
@@ -1,12 +1,0 @@
-export const TIME_REGEX = /^([01]\d|2[0-3]):([0-5]\d)$/;
-export const days = ["MO", "TU", "WE", "TH", "FR", "SA", "SU", "AR", "IA"];
-export const studyProgram = ["T", "I", "S"];
-export const genEdType = ["NO", "SC", "SO", "HU", "IN"];
-export const assessment = ["LETTER", "SU"];
-export const sortBy = ["NAME", "CAPACITY_SUM", "REMAING_SUM"];
-export const sortOrder = ["asc", "desc"];
-export const semester = ["1", "2", "3"];
-export const semesterString = ["FIRST", "SECOND", "SUMMER"];
-export const visible = ["PUB", "PVT"];
-export const reviewStatus = ["PENDING", "APPROVED", "REJECTED"];
-export const vote = ["L", "D"];
