--- apps/core/src/zod_schemas/publicCcarts.schema.ts
@@ -1,11 +1,0 @@
-import { z } from "zod";
-
-export const ImportTimetableBodySchema = z
-  .object({
-    name: z.string().nonempty().optional(),
-  })
-  .strict();
-
-export type ImportTimetableBodySchema = z.output<
-  typeof ImportTimetableBodySchema
->;
